Added project Solution/WinStore/MetroWpf
The C# part's MetroWpf project. This is the native Win8 BookTera
application. This is only a sketch project, it only contains the app's
book-listing parts (the ones one can reach without authentication). The
original goal of this project was to try out the Win8 development, and to
test the WCF real